R = Double again

+++ src/ConCat/Misc.hs
@@ -232,5 +232,5 @@ bottom = error "bottom evaluated"
 typeR :: forall a. Typeable a => TypeRep
 typeR = typeRep (Proxy :: Proxy a)
 
-type R = Float -- Double
+type R = Double -- Float
